+++ localflavor/be/forms.py
@@ -3,8 +3,6 @@
 from django.forms.fields import RegexField, Select
 from django.utils.translation import ugettext_lazy as _
 
-from localflavor.deprecation import DeprecatedPhoneNumberFormFieldMixin
-
 from .be_provinces import PROVINCE_CHOICES
 from .be_regions import REGION_CHOICES
 
@@ -27,47 +25,6 @@ class BEPostalCodeField(RegexField):
 
     def __init__(self, *args, **kwargs):
         super(BEPostalCodeField, self).__init__(r'^[1-9]\d{3}$', *args, **kwargs)
-
-
-class BEPhoneNumberField(RegexField, DeprecatedPhoneNumberFormFieldMixin):
-    """
-    A form field that validates its input as a belgium phone number.
-
-    Landlines have a seven-digit subscriber number and a one-digit area code,
-    while smaller cities have a six-digit subscriber number and a two-digit
-    area code. Cell phones have a six-digit subscriber number and a two-digit
-    area code preceeded by the number 4.
-    0d ddd dd dd, 0d/ddd.dd.dd, 0d.ddd.dd.dd,
-    0dddddddd - dialling a bigger city
-    0dd dd dd dd, 0dd/dd.dd.dd, 0dd.dd.dd.dd,
-    0dddddddd - dialling a smaller city
-    04dd ddd dd dd, 04dd/ddd.dd.dd,
-    04dd.ddd.dd.dd, 04ddddddddd - dialling a mobile number
-
-    .. deprecated:: 1.4
-        Use the django-phonenumber-field_ library instead.
-
-    .. _django-phonenumber-field: https://github.com/stefanfoulis/django-phonenumber-field
-    """
-
-    default_error_messages = {
-        'invalid': _('Enter a valid phone number in one of the formats '
-                     '0x xxx xx xx, 0xx xx xx xx, 04xx xx xx xx, '
-                     '0x/xxx.xx.xx, 0xx/xx.xx.xx, 04xx/xx.xx.xx, '
-                     '0x.xxx.xx.xx, 0xx.xx.xx.xx, 04xx.xx.xx.xx, '
-                     '0xxxxxxxx or 04xxxxxxxx.'),
-    }
-
-    def __init__(self, *args, **kwargs):
-        super(BEPhoneNumberField, self).__init__(r'^[0]\d{1}[/. ]?'
-                                                 r'\d{3}[. ]\d{2}[. ]?'
-                                                 r'\d{2}$|^[0]\d{2}[/. ]?'
-                                                 r'\d{2}[. ]?\d{2}[. ]?'
-                                                 r'\d{2}$|^[0][4]\d{2}[/. ]?'
-                                                 r'\d{2}[. ]?'
-                                                 r'\d{2}[. ]?'
-                                                 r'\d{2}$',
-                                                 *args, **kwargs)
 
 
 class BERegionSelect(Select):