+++ localflavor/dk/forms.py
@@ -6,8 +6,6 @@ from django.core.exceptions import ValidationError
 from django.forms import fields, widgets
 from django.utils.translation import ugettext_lazy as _
 
-from localflavor.deprecation import DeprecatedPhoneNumberFormFieldMixin
-
 from .dk_municipalities import DK_MUNICIPALITIES
 from .dk_postalcodes import DK_POSTALCODES
 
@@ -32,29 +30,4 @@ class DKMunicipalitySelect(widgets.Select):
             choices=DK_MUNICIPALITIES,
             *args,
             **kwargs
-        )
-
-
-class DKPhoneNumberField(fields.RegexField, DeprecatedPhoneNumberFormFieldMixin):
-    """
-    Field with phone number validation.
-
-    Requires a phone number with 8 digits and optional country code.
-
-    .. deprecated:: 1.4
-        Use the django-phonenumber-field_ library instead.
-
-    .. _django-phonenumber-field: https://github.com/stefanfoulis/django-phonenumber-field
-    """
-
-    default_error_messages = {
-        'invalid': _(
-            'A phone number must be 8 digits and may have country code'
-        ),
-    }
-
-    def __init__(self, *args, **kwargs):
-        super(DKPhoneNumberField, self).__init__(
-            r'^(?:\+45)? ?(\d{2}\s?\d{2}\s?\d{2}\s?\d{2})$',
-            *args, **kwargs
         )