+++ localflavor/it/forms.py
@@ -6,17 +6,12 @@ import re
 
 from django.core.validators import EMPTY_VALUES
 from django.forms import ValidationError
-from django.forms.fields import CharField, Field, RegexField, Select
-from django.utils.encoding import force_text
+from django.forms.fields import Field, RegexField, Select
 from django.utils.translation import ugettext_lazy as _
-
-from localflavor.deprecation import DeprecatedPhoneNumberFormFieldMixin
 
 from .it_province import PROVINCE_CHOICES
 from .it_region import REGION_CHOICES, REGION_PROVINCE_CHOICES
 from .util import ssn_validation, vat_number_validation
-
-phone_digits_re = re.compile(r'^(?:\+?39)?((0\d{1,3})(\d{4,8})|(3\d{2})(\d{6,8}))$')
 
 
 class ITZipCodeField(RegexField):
@@ -113,32 +108,3 @@ class ITVatNumberField(Field):
             return vat_number_validation(value)
         except ValueError:
             raise ValidationError(self.error_messages['invalid'])
-
-
-class ITPhoneNumberField(CharField, DeprecatedPhoneNumberFormFieldMixin):
-    """
-    A form field that validates input as an Italian phone number.
-
-    Will strip any +39 country prefix from the number.
-
-    .. versionadded:: 1.1
-
-    .. deprecated:: 1.4
-        Use the django-phonenumber-field_ library instead.
-
-    .. _django-phonenumber-field: https://github.com/stefanfoulis/django-phonenumber-field
-    """
-
-    default_error_messages = {
-        'invalid': _('Enter a valid Italian phone number.'),
-    }
-
-    def clean(self, value):
-        super(ITPhoneNumberField, self).clean(value)
-        if value in self.empty_values:
-            return self.empty_value
-        value = re.sub(r'[^\+\d]', '', force_text(value))
-        m = phone_digits_re.match(value)
-        if m:
-            return '%s %s' % tuple(group for group in m.groups()[1:] if group)
-        raise ValidationError(self.error_messages['invalid'])