+++ tests/test_us/tests.py
@@ -270,24 +270,6 @@ class USLocalFlavorTests(TestCase):
         self.assertHTMLEqual(str(self.form['zip_code']),
                              '<input id="id_zip_code" maxlength="10" name="zip_code" type="text" value="12345" />')
 
-    def test_USPhoneNumberField(self):
-        error_format = ['Phone numbers must be in XXX-XXX-XXXX format.']
-        valid = {
-            '312-555-1212': '312-555-1212',
-            '3125551212': '312-555-1212',
-            '312 555-1212': '312-555-1212',
-            '(312) 555-1212': '312-555-1212',
-            '312 555 1212': '312-555-1212',
-            '312.555.1212': '312-555-1212',
-            '312.555-1212': '312-555-1212',
-            ' (312) 555.1212 ': '312-555-1212',
-        }
-        invalid = {
-            '555-1212': error_format,
-            '312-55-1212': error_format,
-        }
-        self.assertFieldOutput(forms.USPhoneNumberField, valid, invalid)
-
     def test_USStateField(self):
         error_invalid = ['Enter a U.S. state or territory.']
         valid = {